feat: remove BUY and SELL orders
