+++ packages/marketplace/contracts/exchange/libraries/LibFill.sol
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.21;
 
-import {LibOrder, LibMath} from "../../lib-order/LibOrder.sol";
+import {LibOrder} from "../../lib-order/LibOrder.sol";
+import {LibMath} from "./LibMath.sol";
 
 /// @title This library provides `fillOrder` function.
 /// @notice It calculates fill of both orders (part of the Order that can be filled).
@@ -11,18 +12,11 @@ library LibFill {
         uint256 rightValue;
     }
 
-    struct IsMakeFill {
-        bool leftMake;
-        bool rightMake;
-    }
-
     /// @notice Should return filled values
     /// @param leftOrder left order
     /// @param rightOrder right order
     /// @param leftOrderFill current fill of the left order (0 if order is unfilled)
     /// @param rightOrderFill current fill of the right order (0 if order is unfilled)
-    /// @param leftIsMakeFill true if left orders fill is calculated from the make side, false if from the take side
-    /// @param rightIsMakeFill true if right orders fill is calculated from the make side, false if from the take side
     /// @dev We have 3 cases, 1st: left order should be fully filled
     /// @dev 2nd: right order should be fully filled or 3d: both should be fully filled if required values are the same
     /// @return the fill result of both orders
@@ -30,25 +24,28 @@ library LibFill {
         LibOrder.Order calldata leftOrder,
         LibOrder.Order calldata rightOrder,
         uint256 leftOrderFill,
-        uint256 rightOrderFill,
-        bool leftIsMakeFill,
-        bool rightIsMakeFill
+        uint256 rightOrderFill
     ) internal pure returns (FillResult memory) {
-        (uint256 leftMakeValue, uint256 leftTakeValue) = LibOrder.calculateRemaining(
-            leftOrder,
-            leftOrderFill,
-            leftIsMakeFill
-        );
-        (uint256 rightMakeValue, uint256 rightTakeValue) = LibOrder.calculateRemaining(
-            rightOrder,
-            rightOrderFill,
-            rightIsMakeFill
-        );
+        (uint256 leftMakeValue, uint256 leftTakeValue) = calculateRemaining(leftOrder, leftOrderFill);
+        (uint256 rightMakeValue, uint256 rightTakeValue) = calculateRemaining(rightOrder, rightOrderFill);
 
         if (rightTakeValue > leftMakeValue) {
             return fillLeft(leftMakeValue, leftTakeValue, rightOrder.makeAsset.value, rightOrder.takeAsset.value);
         }
         return fillRight(leftOrder.makeAsset.value, leftOrder.takeAsset.value, rightMakeValue, rightTakeValue);
+    }
+
+    /// @notice calculate the remaining fill from orders
+    /// @param order order that we will calculate the remaining fill
+    /// @param fill to be subtracted
+    /// @return makeValue remaining fill from make side
+    /// @return takeValue remaining fill from take side
+    function calculateRemaining(
+        LibOrder.Order calldata order,
+        uint256 fill
+    ) internal pure returns (uint256 makeValue, uint256 takeValue) {
+        takeValue = order.takeAsset.value - fill;
+        makeValue = LibMath.safeGetPartialAmountFloor(order.makeAsset.value, order.takeAsset.value, takeValue);
     }
 
     function fillRight(
