+++ packages/marketplace/contracts/lib-order/LibOrder.md
@@ -14,8 +14,6 @@ This library contains struct `Order` with some functions for this struct:
 - `uint` salt - random number to distinguish different maker's Orders
 - `uint` start - Order can't be matched before this date (optional)
 - `uint` end - Order can't be matched after this date (optional)
-- `bytes4` dataType - type of data, usually hash of some string, e.g.: "v1", "v2" (see more [here](LibOrderData.md))
-- `bytes` data - generic data, can be anything, extendable part of the order (see more [here](LibOrderData.md))
 
 #### Types for EIP-712 signature:
 ```javascript
@@ -36,8 +34,6 @@ const Types = {
 		{name: 'salt', type: 'uint256'},
 		{name: 'start', type: 'uint256'},
 		{name: 'end', type: 'uint256'},
-		{name: 'dataType', type: 'bytes4'},
-		{name: 'data', type: 'bytes'},
 	]
 };
 ```
