+++ packages/marketplace/contracts/lib-part/LibPart.sol
@@ -10,7 +10,7 @@ library LibPart {
     bytes32 public constant TYPE_HASH = keccak256("Part(address account,uint96 value)");
 
     struct Part {
-        address payable account;
+        address account;
         uint96 value;
     }
